+++ R/format_table.R
@@ -12,8 +12,8 @@
 #'   `broom::tidy()`.
 #' @param pretty_names Return "pretty" (i.e. more human readable) parameter
 #'   names.
-#' @param digits,ci_digits,p_digits,rope_digits Number of digits for rounding or
-#'   significant figures. May also be `"signif"` to return significant
+#' @param digits,ci_digits,p_digits,rope_digits,ic_digits Number of digits for
+#'   rounding or significant figures. May also be `"signif"` to return significant
 #'   figures or `"scientific"` to return scientific notation. Control the
 #'   number of digits by adding the value as suffix, e.g. `digits = "scientific4"`
 #'   to have scientific notation with 4 decimal places, or `digits = "signif5"`
@@ -67,6 +67,7 @@ format_table <- function(x,
                          ci_digits = 2,
                          p_digits = 3,
                          rope_digits = 2,
+                         ic_digits = 1,
                          zap_small = FALSE,
                          preserve_attributes = FALSE,
                          exact = TRUE,
@@ -186,7 +187,7 @@ format_table <- function(x,
 
 
   # rename performance columns
-  x <- .format_performance_columns(x, digits, zap_small)
+  x <- .format_performance_columns(x, digits, ic_digits, zap_small)
 
 
   # Format remaining columns
@@ -602,7 +603,7 @@ format_table <- function(x,
 
 
 
-.format_performance_columns <- function(x, digits, zap_small) {
+.format_performance_columns <- function(x, digits, ic_digits, zap_small) {
   if ("R2_adjusted" %in% names(x)) names(x)[names(x) == "R2_adjusted"] <- "R2 (adj.)"
   if ("R2_conditional" %in% names(x)) names(x)[names(x) == "R2_conditional"] <- "R2 (cond.)"
   if ("R2_marginal" %in% names(x)) names(x)[names(x) == "R2_marginal"] <- "R2 (marg.)"
@@ -611,25 +612,58 @@ format_table <- function(x,
   if ("Performance_Score" %in% names(x)) names(x)[names(x) == "Performance_Score"] <- "Performance-Score"
   if ("Wu_Hausman" %in% names(x)) names(x)[names(x) == "Wu_Hausman"] <- "Wu & Hausman"
   if ("p(Wu_Hausman)" %in% names(x)) names(x)[names(x) == "p(Wu_Hausman)"] <- "p(Wu & Hausman)"
-  # add weighted IC to IC columns
+
+  # Formatting if we have IC and IC weight columns ----
+
+  # add weighted IC to IC columns. The next code lines only apply if we have
+  # both the IC and IC weights in the data frame
   all_ics <- list(AIC = c("AIC", "AIC_wt"), BIC = c("BIC", "BIC_wt"),
                   AICc = c("AICc", "AICc_wt"), WAIC = c("WAIC", "WAIC_wt"),
                   LOOIC = c("LOOIC", "LOOIC_wt"))
   for (ic in names(all_ics)) {
     ics <- all_ics[[ic]]
     if (all(ics %in% colnames(x))) {
-      x[[ics[1]]] <- format_value(x[[ics[1]]], digits = digits, zap_small = zap_small)
+      x[[ics[1]]] <- format_value(x[[ics[1]]], digits = ic_digits, zap_small = zap_small)
       x[[ics[2]]] <- format_p(x[[ics[2]]], digits = digits, name = NULL, whitespace = FALSE)
       x[[ics[1]]] <- sprintf("%s (%s)", x[[ics[1]]], x[[ics[2]]])
       x[ics[2]] <- NULL
       names(x)[names(x) == ics[1]] <- sprintf("%s (weights)", ics[1])
     }
   }
-  if ("AIC_wt" %in% names(x)) names(x)[names(x) == "AIC_wt"] <- "AIC weights"
-  if ("BIC_wt" %in% names(x)) names(x)[names(x) == "BIC_wt"] <- "BIC weights"
-  if ("AICc_wt" %in% names(x)) names(x)[names(x) == "AICc_wt"] <- "AICc weights"
-  if ("WAIC_wt" %in% names(x)) names(x)[names(x) == "WAIC_wt"] <- "WAIC weights"
-  if ("LOOIC_wt" %in% names(x)) names(x)[names(x) == "LOOIC_wt"] <- "LOOIC weights"
+
+  # Formatting if we only have IC columns ----
+
+  # if we don't have IC weights, format regular IC columns
+  if ("AIC" %in% names(x)) format_value(x[["AIC"]], digits = ic_digits, zap_small = zap_small)
+  if ("BIC" %in% names(x)) format_value(x[["BIC"]], digits = ic_digits, zap_small = zap_small)
+  if ("AICc" %in% names(x)) format_value(x[["AICc"]], digits = ic_digits, zap_small = zap_small)
+  if ("WAIC" %in% names(x)) format_value(x[["WAIC"]], digits = ic_digits, zap_small = zap_small)
+  if ("LOOIC" %in% names(x)) format_value(x[["LOOIC"]], digits = ic_digits, zap_small = zap_small)
+
+  # Formatting if we only have IC weights ----
+
+  # if we don't have regular ICs, format and rename IC weights
+  if ("AIC_wt" %in% names(x)) {
+    format_p(x[["AIC_wt"]], digits = digits, name = NULL, whitespace = FALSE)
+    names(x)[names(x) == "AIC_wt"] <- "AIC weights"
+  }
+  if ("BIC_wt" %in% names(x)) {
+    format_p(x[["BIC_wt"]], digits = digits, name = NULL, whitespace = FALSE)
+    names(x)[names(x) == "BIC_wt"] <- "BIC weights"
+  }
+  if ("AICc_wt" %in% names(x)) {
+    format_p(x[["AICc_wt"]], digits = digits, name = NULL, whitespace = FALSE)
+    names(x)[names(x) == "AICc_wt"] <- "AICc weights"
+  }
+  if ("WAIC_wt" %in% names(x)) {
+    format_p(x[["WAIC_wt"]], digits = digits, name = NULL, whitespace = FALSE)
+    names(x)[names(x) == "WAIC_wt"] <- "WAIC weights"
+  }
+  if ("LOOIC_wt" %in% names(x)) {
+    format_p(x[["LOOIC_wt"]], digits = digits, name = NULL, whitespace = FALSE)
+    names(x)[names(x) == "LOOIC_wt"] <- "LOOIC weights"
+  }
+
   x
 }